// src/users/dto/update-user.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreateUserDto } from './create-user.dto';

export class UpdateUserDto extends PartialType(CreateUserDto) {
    email?: string | undefined;
    name?: string | undefined;
    password?: string | undefined;
    username?: string | undefined;
    id: string;
    updatedAt: Date;

}